Don't ever quit the server
Quit the server doesn't have any sense under any circustances, so
better if we remove this code.
This is not the cause of any trouble since the actual Quit code should
not be executed (we should always have a server) but we can be certain
that it won't be execuded if we just remove it.