Added KFileItemActionPlugin to send files from dolphin/konqueror
This is one of the most requested features and we have not
implemented it until now because there was no way of doing it right.
Now thanks of KFileItemActionPlugin we can add logic that is
executed before the context menu is shown, so we can know if
there is Bluetooth, if there are known devices, etc.