Use a different Page for each way of launched sendfiles
sendfiles can be launched in tree different ways and now we're
using 3 different pages depending on how it is launched.

-Without UBI or Files
	-SelectDeviceAndFilesPage is used (this is the old DevicePage)
-With UBI
	-SelectFilesPage
-With Files
	-SelectDevicePage

Also this commit changes the way we were getting the files, now the
wizard stores a QStringList instead of the KFileDialog and and therefore
SendFilesJob has been ported to use QStringList instead.