Removed PIN page and put all his logic into Discover
Usually almost all users will be ok by using the automatic mode
and even if they do have to set a manual pin we can add it later
into our database (which is kinda shared with GNOME).

The current interface in DiscoverPage is not final, I know that it
is a little bit cluttered right now.