Start bluedevil-monolithic in onlineMode and kill it on offline.
Don't ask me why I move bluedevil-monolithic launch outside the
modes logic, but the thing is that I did :/ also we rename the
monolithic app from "bluedevil_monolithic" to "bluedevilmonolithic"
so the dbus call should be changed.

I will test this a little bit and if everything is fine I will
backport.