Use the new deviceFound(QVariantMap) in deviceFound
With libbluedevil 1.8 we can't be sure that the device is being
updated correctly due to some internal design errors (that are
fixed in a proper way in master). So, to workaround the problems
a signal has been added which "passes" the whole QVariantMap bluez
emits, so we can be sure that we get fresh data everytime.