First snapshot of "LegacyPairing" page, fully working so far
LegacyPairing is the page that is shown when the device is legacy
and a PIN needs to be introduce (mainly because we can't know the
PIN via database).

At this very moment the page is shown for devices with a PIN in
our database, but this will be fixed shortly.