Only show symlinks as they're the only one created with bluedevil
Now we're preparing the system to have a "non-hidden" root shared
folder, which means that the user will have easy access to the
content and because of this files can be shared bypassing bluedevil
which is actually fine, but in our system we've to show only the
files shared specifically from bluedevil.

The solution of showing only symlinks is not perfect, yet it is
enough since we don't really want to have to track the shared
files, save them in a file etc.