Added knownDevices to the KDED, returning the list of known devices
Copy/Pasted from the doc:
This slot will return a list of devices made of: configured and discovered devices.
Going deeper, the first time that this slot is called a discovery of X seconds will start
Then if this slot is consulted again it will return configured and discovered device. Once
the discovery ends it won't start a new discovery until N seconds pass.