Show always the knownDevices while running a discovery at the background
This fixes one of the most hated behaviors of kio_bluetooth, before this
commit we were scanning always for new devices so in order to use your
device via kio_bluetooth it had to be in discoverable mode (and of
of course you had to wait a bit).
Now, we're showing always the knownDevices first, and start a
discovery each 20s. If a folder is acceed the discovery is stopped.

BUG: 253810
FIXED-IN: 1.2