Make 'Enter' and 'Return' work to accept the dialog
Users need to be able to hit 'Enter' or 'Return' after entering the pin
so that they don't have to move their hands off the keyboard. An even
better reason to do this is because 95% of the people are going to hit
'Enter' and expect the dialog to just work.