Added support for PANU network
Well, this may be a little bit hackish because of all the repeated code
but with the current plugin system I'm afraid it is the only way.

For BlueDevil 1.3 I want to fix the situation because with the current
system I can't improve the Wizard workflow by removing the last step.