Replaced AgentListener (QThread) by delayed DBus messages
This thread has been causing many crashes for some people, and
as far as I remember all the backtraces show weird things not really
within BlueDevilcode, and of course none of the BlueDevil developers has
been able to reproduce the bug. So the best/easy/simpler way of fixing
this is by removing the Thread which shouldn't be there in the first
place.

BUG: 262587
BUG: 275783
BUG: 262041
BUG: 283365