Revert "Make BlueDevil system tray icon Passive (i. e. hide it) if adapters are turned off."
If the Bluetooth hardware is turned off by software, this probably means
that the computer where BlueDevil is running DOES NOT HAVE a hardware
switch so the userd had to use the software switch provided in the
monolithic. I don't think it improve usability by hidding the bluetooth
icon in thar case, since the user will have to do 3 clicks instead
of 2 (apart from discover that the Bluetooth icon is hidden).

I'm going to bring the topic to plasma-devel and see what other
people think, if the rest thinks like your then I will revert my
revert :p

This reverts commit 65e04a744248b9b598aecec8939cebaa5d3d15d3.