Set a new Object as parent of BlueZAgent
You can't register a child of a parent exposed with ExportChildObjects, if
you do so the app may crash.

for not making this patch too big I'm using the parent() reference from
the BlueAgent object itself to delete the parent and this will delete
the agent.  I will improve it in BlueDevil 2.0

Versions minor to 1.3 are not affected by this bug since they use qApp
as a parent instead of BlueDevilDAemon.

BUG: 284052