Check if we have permission to save the files in the destination folder
First we check if the destination folder exists and is writeable.
If it doesn't exists we try to create it.

If we can't use the configure destination folder, we fallback to
DocumentsLocation also checking if it doesn't exists and creating it
en case it does exists.

BUG:294041
(cherry picked from commit 21eb2f540a1e20fb9e52a4141f83f684fd287247)

Conflicts:

	src/daemon/helpers/filereceiver/openobex/server.cpp