Check if there is an usable adapter before going to onlineMode
An adapter can be added but it might no be usable (not powered) so we
have to be certain that there is an usableAdapter before going into
onlineMode.

In the future We should connect only to usableAdapterChanged to make
the code waay simpler, but since we are close to release let's not
modify too much code.

CCBUG: 314356