Show all devices known in the wizard
This is a small change of behavior compared to the bluez4 bluedevil
version, but I think it is qutie handy since it will allow the wizard
to stablish connection with profiles of already created devices.

Also, this makes porting to bluez5 easier since bluez5 is designed to be
used this way, wich I think is the right way.