New bluez5 port PoC for kio_obexftp
Re-writing kio-obexftp to bluez5 is resulting in quite complex code
given that now doing it in a synchronous way is NOT an option.

Currently it looks like this:

kio-------StablishConnection
 .                |
 .               SYNC
 .                |
 .              KDED-------CreateSession
 .                                |
 .                              ASYNC
 .                                |
 .              KDED<-----org.bluez.obex
 .               |
------------------

KIO uses dbus to talk with KDED in a synchronous way, KDED keeps the
message and calls CreateSession to org.bluez.obex ASYNC, when it
finishes replies to kio (that called it synchronously) with the values
or errors.