First draft to support Kio::stat
Stat is resulting to be pretty much like listDir so we will have to
figure out a way of removing the code duplication.

But for now, it works.

The mapStat cache is again being used but I'm not paying too much
attention to it, it might be broken.

Finally changeCurrentFolder has been added.