Remove ServicePage
With bluez5 we don't need the user to choose anything bluez5 knows
what to do.

We are missing some things like being able to send files or browse
it from the wizard, which is rather pointless since you can send
files and pairing will start if required anyway. Same for
browsing files.