Show always the KStatusNotifier (Status == Active)
The KDED will kill the monolithic in case there are no
adapters so in that case we won't polute the systemtray
with the Bluetooth icon.

In the case there there are Adapters but they are not
powered we want to show the systemtray so it can be
powered on again.

BUG: 319403