Fix possible incorrect filename when receiving files
All incoming file transfers are temporarily downloaded into temp folder
($CACHE/obexd) and then moved to the target folder.
When there is already a file with the same name in temp folder, unique
name is created (by appending a number to original name).

Make sure this modified filename is only used internally, and never
displayed to user (in notification).