[applet] Show toolbar unless Bluetooth hardware is absent
Summary:
While testing other bluedevil patches, I was finding that the toolbar showing and hiding
itself when bluetooth is turned on and off was feeling a bit weird to me. This patch
makes the toolbar always visible as long as a Bluetooth adapter is present; effectively
this means that the toolbar no longer disappears when bluetooth is disabled. Instead,
the toolbar remains visible but its controls are all disabled.

If nobody likes this, I can drop it, but wanted to present the idea. If we like this,
I would propose doing the same for other applets with toolbars so that they aren't
appearing and disappearing so much.

Test Plan: {F8255623}

Reviewers: #vdg, niccolove, manueljlin

Reviewed By: #vdg, manueljlin

Subscribers: broulik, plasma-devel

Tags: #plasma

Maniphest Tasks: T10470

Differential Revision: https://phabricator.kde.org/D29132