New bluetooth KCM
This is a total rewrite of the bluetooth KCMs with current technology and visual style. It combines the three individual KCMs into a single one using multiple pages.

Functionality-wise it is mostly equivalent, except for two differences.

1) The new KCM does not allow setting an adapter visibility timeout

2) There is no option to not receive files any more. It is somewhat redundant with the default setting of not receiving them automatically.

Adding new devices is not yet ported, instead this just opens the wizard just like the old KCM. We sure want to have it properly integrated at some point, but it's a rather complex piece in itself and the diff is already huge.

Fixes #1