Show "Enable bluetooth" only when Manager is operational
The properties are fetched asynchronously, so when opening the KCM it would always briefly
flash the "Enable bluetooth" hint until the list is loaded.

Also check whether the manager itself is operational and don't show it when this isn't the case.

Ideally, we'd had a way to check whether BluezQt is fecthing properties and show a BusyIndicator
and also know when it fails to show a respective error. However, this is a simple fix
that makes the UX slightly more polished.

(cherry picked from commit b53d7356f18dca5be4d5a22ef439087c4df12649)