Show only paired devices in KCM and applet
Currently only paired/known devices are shown initially. Once the user starts the
wizard we enter discover mode and all newly found devices are added to
the model and thus the applet and KCM. When the KCM is closed and opened
again only the paired devices.

To avoid this filter out unpaired devices in the ProxyModel that is used
by the applet and KCM.

BUG: 422383
BUG: 416322