Confirm deleting a bluetooth device pairing.
The deletion of a pairing was instant without undo and no confirmation,
made worse by the fact that the small minus icon in breeze made it not
seem dangerous.

This adds a confirmation dialog in order to inform the user what happens
and allow them to cancel the deletion of the pairing.

Also: Replace delete with red cross icon