Remove applet list item animations
This was always the indended appearance, but it was being overridden by
a bug in the ExpandableListItem itself. Let's remove the animations here
so that the bug in question can be fixed while preserving the intended
visual effect.

(cherry picked from commit d8b6b0359540587d7f7e680d25a34ad8ae9abc55)