Show connected but non-paired devices
Devices like Wiimotes and DualShock 3's use a custom method of
connecting. This causes those devices to not actually be paired like
usual but still be capable of connecting to the system.
These devices however never showed up in either the KCM or the applet as
they only showed paired devices, even if the devices were connected.

BUG: 432715