applet: introduce a brief animation for section height estimates
section delegates changing height is a bit buggy.
as a simple bandaid try to delay the height change a bit via a behavior.
NOTE: this doesn't actually animate, it's purely there for the delay
aspect

huge hack but it gives the view a better chance of figuring out a layout
that works from my testing

QTBUG-29148
BUG: 438610