applet: Simplify and fix device change handler in delegate
Actually this code path is not being triggered at all even with
ListView.reuseItems because the sort filter model either preserves the
delegates or recreates them but not reuses across different devices.

Either way, that `status === Component.Ready` condition did not make any
sense for a statically declared Component, and we should never reset
the ListView.currentIndex in delegates if the index is not ours.

Fixes #5