kcm: Hide or disable Connect buttons for blocked devices
The upstream BlueZ project allows blocking devices which is inhibits
their ability to initiate connections to the device. However, for some
reasons (probably due to a bug or an assumption about the async nature)
it still allows attempting to connect to it, but will actually connect
and disconnect twice before bailing out with an error like this:

    org.bluez.Error.Failed br-connection-aborted-by-remote

which doesn't even sound correct because the remove never connects --
evidences by testing on the headphones which usually play "Bluetooth
connected" voice line whenever they actually connect, which in this
case they don't do.

So to solve this we should prevent users from attempting to connect to
their blocked devices by disabling the Connect button or even hiding
the entries completely. Hiding in the applet makes sense because
applets are designed for quick access and not complete system
management.

See also:

- https://github.com/bluez/bluez/blob/52bda9d45572ab4629bf2f686616c0398f489dad/doc/settings-storage.txt#L290
- https://github.com/bluez/bluez/blob/52bda9d45572ab4629bf2f686616c0398f489dad/doc/org.bluez.Device.rst#L221
- commit a527e907f5c2e982ecbd3180e199dc56f1dc3fdc for applet.