applet: Factor the state out of delegates to a shared proxy model
None of this belongs to the delegate's UI code. If model rows are
shuffled around or delegates are recreated by the ListView for other
reasons -- the state will be lost, or worse: retained with a new
incorrect index/device. Thus, all and any state should be stored in the
upstream models or middleware proxy models. To keep the state in sync
between all instances of the Bluetooth applet, the new proxy model is
exposed as a singleton, although it needed some trickery with .qml file
to make it work, since we can't access BlueZ-Qt QML module's types from
C++ in Bluedevil.

Error handling logic is moved out of the delegates as well. The proxy
model now submits a KNotification itself, because, again, the delegate
might be scrolled out of view and get destroyed so it won't be able to
run the code safety for no good reason.