kcm: only display blocked tooltip if device blocked
On a disabled "Connect" button in the device page of the bluedevil kcm,
a helpful tooltip informs the user that they need to unblock the device
to connect. This tooltip does not take into account that the button may
be disabled for other reasons than the device being blocked, such as the
device already being in the process of connecting, such as immediately
after pressing the button. And in this case, the tooltip is just
confusing, rather than helpful.

This change replaces the condition for displaying the tooltip, only
showing it when the text actually applies, namely when the device
is blocked.


(cherry picked from commit a6295326fe04acb18199af0d70d53929915c1291)

Co-authored-by: Christoph Wolk <christoph.wolk@gmx.de>