kcm: use switches for instant-apply options
The HIG recommend the use of switches instead of checkboxes for settings
that take effect immediately. The page for devices and the settings page
of the Bluetooth kcm have such settings (e.g. "Trusted" or "Visible")
that don't require pressing apply, but are nonetheless checkboxes rather
than switches.

This change replaces these checkboxes with switches.