applet: Hide devices if bluetooth adapter is powered off
The bluetooth daemon can report devices if the adapter is powered off.
In this case, the applet is going to display those devices but if you
click the connect button, the connect operation will fail for obvious
reasons. It is misleading, it is great if you are able to spot immediately
that the enabled slider is in the off position, but it is likely that
you'll miss it.

Currently, the "bluetooth is disabled" placeholder message is wired to
Manager.bluetoothBlocked, which is different from the powered state. If
the bluetooth adapter is powered off, it is not necessarily blocked.

With this change, the "bluetooth is disabled" label is synchronized
to Manager.bluetoothOperational, similar to the enable slider. The
devices will be shown if bluetooth is not blocked, and the adapter is
powered on.