applet: fix standalone widget toolbar
The standalone widget toolbar was not adapted for the changes made for
the badge interface; it still accesses the internal configure action
which was removed to place it in the menu in the system tray, breaking
that button, and it does not expose the button to toggle the badge
display.

This change reproduced the current system tray buttons also for the
standalone widget toolbar. Along the way, we switch to setting icononly
display style instead of having empty text for the buttons.