DevicesStateProxyModel: Split disconnecting into a dedicated state
While the method "registerPendingCall" sounds generic, it's really only
about the connecting state. Unfortunately, a device can be considered
connected before the pending call has finished. This causes the UI to
briefly display a "Disconnecting" in the UI which is confusing.

By properly tracking connecting and disconnecting calls we can ensure
that the device state displayed matches expectations.

BUG: 493245