wizard: allow filtering device list
This change makes the list of scanned devices more manageable. It allows
to filter out devices that don't have a human-readable name (yet). For
the user it's hard to know what these devices are, and bluetooth devices
that are interesting to pair usually offer their name when pairing.

The filter uses the remote name (in bluedevil terms) or the alias (in
bluez terms), if the remote name is empty, devices will just have an
address (determined by bluez).

The device discovery widget now has a checkbox (defaulting to off) which
means "just devices with type and real name", if the user can't find
their device, the checkbox can be checked and all found devices will be
shown.

The default is now in line with how iOS and Android do it (though these
OSes don't allow showing all devices).

Rationale: in our office, the device list can be very long, with 50+
devices found and most of them not useful at all.

I'm a little unclear if and how the "unknown" devices are useful (they
aren't to me, but that doesn't mean that users won't miss them), so I
made them still accessible as to avoid causing regressions.

Signed-off-by: Sebastian Kügler <sebas@kde.org>