Rework adapter state save/restore
Currently we save the adapter power state, but only when rfkill is not blocked.

When the adapter is saved as powered on, and then we block bluetooth, the saved
state is never changed.

On next startup we then power the adapter on. Apparently on some systems that
results in the rfkill being undone and enables bluetooth when it shouldn't be.

To avoid this

1) Clear the adapter state list on save when bluetooth is blocked
2) Only changed the powered state for adapters where we have remembered a
power state, instead of defaulting to powering it on

BUG: 497353

BUG: 507836